ost/requirements.txt
requests
python-whois
colorama